Remove Home. Rename analytics to history.
