--- app/src/main/java/hiof/gruppe15/treningsappen/ui/component/home/Home.kt
@@ -1,34 +1,0 @@
-package hiof.gruppe15.treningsappen.ui.component.home
-
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Button
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.navigation.NavController
-import hiof.gruppe15.treningsappen.ui.component.navigation.AppScaffold
-import hiof.gruppe15.treningsappen.ui.component.navigation.Screen
-
-@Composable
-fun HomeScreen(navController: NavController) {
-    AppScaffold(navController = navController, title = "Home") {
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(it),
-            verticalArrangement = Arrangement.Top,
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-            Button(onClick = {
-                navController.navigate(Screen.Testing.route)
-            }) {
-                Text("Testing")
-            }
-        }
-    }
-}
-
