+++ app/src/main/java/hiof/gruppe15/treningsappen/ui/component/navigation/NavGraph.kt
@@ -6,9 +6,7 @@ import androidx.navigation.NavType
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.navArgument
-import hiof.gruppe15.treningsappen.ui.component.analytics.AnalyticsScreen
-import hiof.gruppe15.treningsappen.ui.component.home.HomeScreen
-import hiof.gruppe15.treningsappen.ui.component.home.TestingScreen
+import hiof.gruppe15.treningsappen.ui.component.history.HistoryScreen
 import hiof.gruppe15.treningsappen.ui.component.login.ForgotPasswordScreen
 import hiof.gruppe15.treningsappen.ui.component.login.LoginScreen
 import hiof.gruppe15.treningsappen.ui.component.login.RegisterScreen
@@ -24,7 +22,6 @@ sealed class Screen(val route: String) {
     object Login : Screen("login")
     object Register : Screen("register")
     object ForgotPassword : Screen("forgotPassword")
-    object Home : Screen("home")
     object Routines : Screen(route = "workoutPlan")
     object RoutineDetails : Screen(route = "routineDetails/{routineId}") {
         fun createRoute(routineId: String) = "routineDetails/$routineId"
@@ -32,15 +29,13 @@ sealed class Screen(val route: String) {
     object WorkoutSession : Screen(route = "workoutSession")
     object CreateNewRoutine : Screen("createNewRoutine")
     object SaveNewRoutine : Screen("saveNewRoutine")
-    object Analytics : Screen(route = "analytics")
+    object History : Screen(route = "history")
     object Profile : Screen("profile")
-    object Testing : Screen("testing")
 }
 
 sealed class ScreenCategory(private val baseRoutes: List<String>) {
-    object Home : ScreenCategory(listOf("home", "testing"))
     object Routines : ScreenCategory(listOf("workoutPlan", "workoutSession", "createNewRoutine", "saveNewRoutine", "routineDetails"))
-    object Analytics : ScreenCategory(listOf("analytics"))
+    object History : ScreenCategory(listOf("history"))
     object Profile : ScreenCategory(listOf("profile"))
 
     fun includes(route: String?): Boolean {
@@ -62,9 +57,6 @@ fun NavGraph(
         composable(Screen.Register.route) { RegisterScreen(navController) }
         composable(Screen.ForgotPassword.route) { ForgotPasswordScreen(navController) }
 
-        composable(Screen.Home.route) { HomeScreen(navController) }
-        composable(Screen.Testing.route) { TestingScreen(navController) }
-
         composable(Screen.Routines.route) { RoutineScreen(navController, sharedViewModel) }
         composable(
             route = Screen.RoutineDetails.route,
@@ -79,7 +71,7 @@ fun NavGraph(
         composable(Screen.CreateNewRoutine.route) { CreateRoutineScreen(navController, sharedViewModel) }
         composable(Screen.SaveNewRoutine.route) { SaveRoutineScreen(navController, sharedViewModel) }
 
-        composable(Screen.Analytics.route) { AnalyticsScreen(navController) }
+        composable(Screen.History.route) { HistoryScreen(navController) }
 
         composable(Screen.Profile.route) { ProfileScreen(navController, sharedViewModel) }
     }
