+++ app/src/main/res/raw/exercises.json
@@ -55,7 +55,7 @@
     "name": "Calf Raises",
     "description": "Raises to strengthen calves.",
     "steps": [
-      "Stand on edge of step.",
+      "Stand on edge of step",
       "Raise heels above step.",
       "Lower heels below step level."
     ],
